Split slack_lead_qualifier.py into files
